# docs/swagger.yaml
basePath: /
definitions:
  models.Announcement:
    properties:
      author:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 公告作者
      authorId:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      guildId:
        type: integer
      id:
        type: integer
      title:
        type: string
    type: object
...
  models.User:
    properties:
      avatar:
        description: 头像文件路径(MinIO object name)
        type: string
      banner:
        description: 横幅背景图片路径(MinIO object name)
        type: string
      bannerColor:
        description: '横幅背景颜色(如 #5865F2)'
        type: string
      bio:
        description: 个人简介
        type: string
      createdAt:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      id:
        type: integer
      isBot:
        description: 机器人账号标记
        type: boolean
      isPrivate:
        description: 私密账号，默认公开
        type: boolean
      name:
        type: string
      nickname:
        description: 扩展字段（不存储在数据库，仅用于API响应）
        type: string
      status:
        type: string
      token:
        type: string
    type: object
...
info:
  contact: {}
  description: Simple chat server (Gin + Gorm)
  title: Bubble Chat API
  version: 0.1.0
paths:
  /api/announcements/{id}:
    delete:
      consumes:
      - application/json
      description: 删除公告（需要管理员权限）
      parameters:
      - description: Announcement ID
        in: path
        name: id
        required: true
        type: integer
      - description: Guild ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 公告不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Delete announcement
      tags:
      - announcements
      - announcements
    get:
      consumes:
      - application/json
      description: 获取公告详情
      parameters:
      - description: Announcement ID
        in: path
        name: id
        required: true
        type: integer
      - description: Guild ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Announcement'
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 公告不存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Get announcement
      tags:
      - announcements
      - announcements
  /api/applications:
    post:
      consumes:
      - application/json
      parameters:
      - description: Application payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create application (friend or guild join)
      tags:
      - me
  /api/applications/{id}:
    put:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: approved or rejected
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Approve/Reject application
      tags:
      - me
  /api/blacklist:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: List blacklist
      tags:
      - blacklist
    post:
      consumes:
      - application/json
      parameters:
      - description: User to block
        in: body
        name: body
        required: true
        schema:
          properties:
            userId:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Add to blacklist
      tags:
      - blacklist
  /api/blacklist/{userId}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Remove from blacklist
      tags:
      - blacklist
...
  /api/bot/v1/pub/search:
    get:
      parameters:
      - description: 搜索关键词 (>=1 字符)
        in: query
        name: q
        required: true
        type: string
      - description: Limit (默认20, 最大100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search robots by name
      tags:
      - robot-public
  /api/channels:
    get:
      consumes:
      - application/json
      description: 列出某服务器的全部频道（用户需具备查看权限）
      parameters:
      - description: Guild ID
        in: query
        name: guildId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List channels
      tags:
      - channels
    post:
      consumes:
      - application/json
      description: 创建频道
      parameters:
      - description: '{guildId,name,type,parentId,categoryId}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create channel
      tags:
      - channels
  /api/channels/{id}:
    delete:
      consumes:
      - application/json
      description: 软删除频道及其子频道（需要 PermManageChannels 权限）
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 频道不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete channel (soft delete)
      tags:
      - channels
  /api/channels/{id}/category:
    put:
      consumes:
      - application/json
      description: 设置频道所属分类
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      - description: '{categoryId?: number|null}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 频道不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set channel category
      tags:
      - channels
...
  /api/dm/open:
    post:
      consumes:
      - application/json
      parameters:
      - description: '{toUserId}'
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Open or get DM thread
      tags:
      - dm
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"